commodities/requirements.txt
scikit-learn
tensorflow
numpy
pandas
alpaca_trade_api
requests
streamlit
alphavantage
azure-storage-blob
openai
scipy